Day 10: Add a "check all inbetween"
This is so rad. We were supposed to try it on our own and I didn't
write the code but I mapped out how I thought I should to it. While
I was in range, my solution was way more complex. I thought I needed
to make arrays to determine order, but flipping a variable was so
much better.